D: C++ audit fixes — int-overflow in server pooling, empty-output guard, sampler leak (pass 2)
From the adversarial C++ re-audit (AUDIT-2026-07-30-D-CPP-RETRY.md). All 13 prior
fixes re-verified present; self-test 17/17. New fixes:

- P1: server-context.cpp:2512,2521-2522 — int32 pointer-offset arithmetic overflowed
  for large ctx·embd (t*n_embd computed in int before the pointer add → wrapped
  pointer / OOB read). The pool=none path at :2492 already cast to size_t; the
  skip_mean and last-token paths now match. Cast to size_t.
- P1: hs-extract-batch.cpp parse_args — reject --token-skip >= --generate. Without
  it the gate 'gen_step >= token_skip' never passes, every layer's gen_count stays 0,
  output.bin gets a header with no data, and the run exits 0 — a 239K-prompt run
  wasted silently. Now exits 1 with a clear error.
- P2: hs-extract-batch.cpp — sampler chain leaked on all 3 generation error paths
  (no-logits, decode-fail, extract-fail) which returned before the cleanup at :1911.
  Free the sampler before each early return.

Verified: full Vulkan rebuild clean, self-test 17/17, token-skip guard exits 1.

+++ tools/server/server-context.cpp
@@ -2501,7 +2501,10 @@ private:
                 int32_t count = n_hs_tokens - start;
                 vec.resize(n_embd, 0.0f);
                 for (int32_t t = start; t < n_hs_tokens; t++) {
-                    const float * tok = hs + t * n_embd;
+                    // size_t cast: t*n_embd is computed in int32 and overflows for
+                    // large ctx·embd (the pool=none path at :2492 already casts;
+                    // this path must match or it reads from a wrapped pointer).
+                    const float * tok = hs + (size_t)t * n_embd;
                     CR_SIMD
                     for (int d = 0; d < n_embd; d++) vec[d] += tok[d];
                 }
@@ -2510,8 +2513,8 @@ private:
                 for (int d = 0; d < n_embd; d++) vec[d] *= inv;
             } else {
                 // Default: last token's hidden state
-                vec.assign(hs + (n_hs_tokens - 1) * n_embd,
-                           hs + n_hs_tokens * n_embd);
+                vec.assign(hs + (size_t)(n_hs_tokens - 1) * n_embd,
+                           hs + (size_t)n_hs_tokens * n_embd);
             }
 
             if (slot.task->params.hidden_normalize) {
